Update template.rs
When using the CLI, it tells me the example is:
EXAMPLES:
    trinsic template create --name 'My Credential' --fields-data '{\"firstName\":{}}'

Using that command directly gives me a "serialization error"

This fix removes the extra backslashes around firstName. The intent is to enable people to simply copy/paste the example command and have a successful call without getting a serialization error.