Update migrate.md (#1001)
What is the value of the "who should consider migration" section? shouldn't it be everyone?

On wallets, it's important we get messaging right. We should call our old wallets "custodian wallets" and new wallets "hybrid wallets" IMO

I added a "without the same privacy guarantees" to the revocation section to make sure it's clear the trade-off we made going with status list